pyproject.toml
[tool.poetry]
name = "betting_bot"
version = "0.1.0"
description = ""
authors = ["rfoucart <rfoucart@ensc.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.11"
pandas = "^2.1.1"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
requests = "^2.32.2"


[tool.pytest.ini_options]
pythonpath = [
    ".",
    "./src",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
